StackOverflow to the rescue! Adds compare_type=True to the migrate constructor.

+++ models.py
@@ -5,7 +5,7 @@ class Path(db.Model):
     __tablename__ = 'paths'
     id = db.Column(db.Integer, primary_key=True)
     svg = db.Column(db.Text, nullable=False)
-    zoom = db.Column(db.Float, nullable=False)
+    zoom = db.Column(db.Integer, nullable=False)
     lat = db.Column(db.Float, nullable=False)
     lng = db.Column(db.Float, nullable=False)
